Pull object diffing into helpers
