+++ src/services/helpers.js
@@ -17,3 +17,16 @@ export const successToast = (toast, detail) => {
         life: toastDuration,
     })
 }
+
+export const objectDiff = (obj1, obj2) => {
+    let res = {};
+	Object.keys(obj1).forEach((key) => {
+		if (typeof obj1[key] === "object") {
+            let deepCompare = objectDiff(obj1[key], obj2[key]);
+            if(Object.keys(deepCompare).length > 0) res[key] = deepCompare;
+		} else if (obj1[key] != obj2[key]) {
+			res[key] = obj1[key];
+		}
+	});
+	return res;
+};
